+++ pkg/cmd/root.go
@@ -59,6 +59,7 @@ func setDefaultConfig() {
 	viper.SetDefault("maia.auth_driver", "keystone")
 	viper.SetDefault("maia.storage_driver", "prometheus")
 	viper.SetDefault("maia.label_value_ttl", "1h")
+	viper.SetDefault("maia.label_value_for_global_visibility", "")
 	viper.SetDefault("keystone.token_cache_time", "900s")
 	viper.SetDefault("keystone.roles", "monitoring_viewer,monitoring_admin")
 	viper.SetDefault("keystone.default_user_domain_name", "Default")